// resources/js/api/generated/models/postApiV1ServersServeridStacksStacknameFilesChownBody.ts
/**
 * Generated by orval v8.0.2 🍺
 * Do not edit manually.
 * Berth API
 * Berth: Opinionated docker compose stack management API
 * OpenAPI spec version: 1.0.0
 */

export type PostApiV1ServersServeridStacksStacknameFilesChownBody = {
  /** @minimum 0 */
  group_id?: number;
  /** @minimum 0 */
  owner_id?: number;
  path: string;
  recursive?: boolean;
};